fix(desktop): exclude archived sessions from workspace auto-select
The auto-select added in 50985fe picked the most-recently-updated session
unconditionally. Archive is a localStorage-only filter, so the freshly
loaded `sessions` array still contains every archived session — and an
archived session has a high updatedAt by construction (the user touched
it right before archiving). Net effect: switching workspaces could
resurrect the session you just archived, point the worktree picker at
its on-disk path, and run agents against the wrong branch.

Read the archived map via a shared helper and filter it out before
picking the most-recent session. The helper now lives in
shared/lib/storage-keys.ts so both the store action and the sidebar
hook use the same source.

Drops the duplicate readArchivedSet defined inside WorkspacesSidebar.
