+++ apps/desktop/src/features/workspace/components/WorkspacesSidebar/index.tsx
@@ -79,7 +79,7 @@ import {
 } from '../../../../features/session/components/AgentMetricsBlock';
 import { formatError } from '../../../../shared/lib/errors';
 import { useThemeStore } from '../../../../shared/lib/theme';
-import { STORAGE_KEYS } from '../../../../shared/lib/storage-keys';
+import { STORAGE_KEYS, readArchivedSessionsFromStorage } from '../../../../shared/lib/storage-keys';
 import { WorkspaceSelect } from '../WorkspaceSelect';
 import { SessionActivityBar } from '../SessionActivityBar';
 import { SessionDetailPanel, SessionFilesTouchedFooter } from '../SessionDetailPanel';
@@ -366,22 +366,6 @@ function NoWorkspaceEmpty({ onAddWorkspace }: { onAddWorkspace: () => void }) {
 
 const ARCHIVED_KEY = STORAGE_KEYS.archivedTasks;
 
-function readArchivedSet(): Record<string, true> {
-  try {
-    const raw = localStorage.getItem(ARCHIVED_KEY);
-    if (!raw) return {};
-    const parsed = JSON.parse(raw) as unknown;
-    if (parsed && typeof parsed === 'object') {
-      const out: Record<string, true> = {};
-      for (const k of Object.keys(parsed as Record<string, unknown>)) out[k] = true;
-      return out;
-    }
-  } catch {
-    // ignore
-  }
-  return {};
-}
-
 function writeArchivedSet(map: Record<string, true>): void {
   try {
     localStorage.setItem(ARCHIVED_KEY, JSON.stringify(map));
@@ -395,7 +379,7 @@ function useArchivedSessions(): [
   (id: SessionId) => void,
   (id: SessionId) => void,
 ] {
-  const [map, setMap] = useState<Record<string, true>>(() => readArchivedSet());
+  const [map, setMap] = useState<Record<string, true>>(() => readArchivedSessionsFromStorage());
   // Stable refs so memoized SessionRow downstream doesn't invalidate.
   const archive = useCallback((id: SessionId) => {
     setMap((prev) => {
